FIX: Remove old code that was moved elsewhere
